add givbacks and qf badges to stellar
