+++ lang/ca.json
@@ -728,6 +728,7 @@
 	"label.photo_by": "Foto per",
 	"label.please_confirm_your_email": "Si us plau, confirma el teu correu electrònic",
 	"label.please_connect_your_wallet": "Si us plau, connecta la teva cartera",
+	"label.please_connect_your_wallet_to_get_matched": "Connecta la teva cartera per igualar la teva donació.",
 	"label.please_contact_support_team": "Si us plau, contacteu l'equip de suport.",
 	"label.please_do_not_enter_exchange_deposit": "Si us plau, NO introduïu una adreça de dipòsit d'intercanvi, o els vostres fons podrien perdre's! Utilitzeu un compte que controleu en aquesta xarxa. Recomanem utilitzar Metamask.",
 	"label.please_enter_full_link": "Introdueix el vincle complet",
