+++ lang/en.json
@@ -730,6 +730,7 @@
 	"label.photo_by": "Photo by",
 	"label.please_confirm_your_email": "Please confirm your email",
 	"label.please_connect_your_wallet": "Please connect your wallet",
+	"label.please_connect_your_wallet_to_get_matched": "Connect your wallet to get your donation matched.",
 	"label.please_contact_support_team": "Please contact support team.",
 	"label.please_do_not_enter_exchange_deposit": "Please DO NOT enter an exchange deposit address, or your funds maybe lost! Use an account you control on this network. We recommend using Metamask.",
 	"label.please_enter_full_link": "Please enter full link",
