+++ src/components/views/donate/OneTime/SelectTokenModal/QRCodeDonation/CopyConatainer.tsx
