+++ src/components/views/donate/OneTime/SelectTokenModal/QRCodeDonation/QRDonationCardContent.tsx
