+++ src/components/views/donate/OneTime/SelectTokenModal/QRCodeDonation/QRDonationDetails.tsx
