+++ src/components/views/donate/OneTime/SelectTokenModal/TokenInfo.tsx
